In VaultToGit.py, updated the logic to clear the working dir to use 'git rm -r .' instead of 'git rm .' and, critically, to not do this on the final loop iteration (i.e. leave the Vault tips as the final contents of the new git repo).
In strip_xml_entities.py, added logic to strip nested quotation marks from the
XML attributes. There are certainly better ways of doing this. This is not
"XML-aware" logic and is therefore brittle, but I believe that to be okay since
the formatting of the attributes as they're pulled from Vault should be brittle
as well. It should also fail safe insofar as it will not modify a line at all
if it does not exactly match the format it expects.